# docs/apidocs/taskapi_openapi_v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Thor.TaskApi | v1",
    "version": "1.0"
  },
  "paths": {
    "/v1/tasks": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{taskId}/heartbeat": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/tasks/{taskId}/status": {
      "put": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskStatusRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskStatusRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/tasks/{taskId}/progress": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskProgressRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskProgressRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TaskProgressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskProgressResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskProgressResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskProgressResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{taskId}/settings": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSettingsResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSettingsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSettingsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/uploads/presigned-url": {
      "post": {
        "tags": [
          "Uploads"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignedUploadRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignedUploadRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PresignedUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUploadResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUploadResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUploadResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ConnectorTaskResponse": {
        "required": [
          "scanId",
          "scanConfigId",
          "taskId",
          "connectorType",
          "authMethod",
          "sourceId"
        ],
        "type": "object",
        "properties": {
          "scanId": {
            "type": "string",
            "format": "uuid"
          },
          "scanConfigId": {
            "type": "string",
            "format": "uuid"
          },
          "taskId": {
            "type": "string",
            "format": "uuid"
          },
          "connectorType": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int16"
          },
          "authMethod": {
            "type": "string",
            "format": "uuid"
          },
          "sourceId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "JsonElement": { },
      "PresignedUploadRequest": {
        "required": [
          "taskId",
          "sourceId",
          "contentType"
        ],
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string",
            "format": "uuid"
          },
          "sourceId": {
            "type": "string",
            "format": "uuid"
          },
          "contentType": {
            "type": "string"
          }
        }
      },
      "PresignedUploadResponse": {
        "required": [
          "uploadUrl",
          "objectKey",
          "expiresAtUtc"
        ],
        "type": "object",
        "properties": {
          "uploadUrl": {
            "type": "string"
          },
          "objectKey": {
            "type": "string"
          },
          "expiresAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TaskListResponse": {
        "required": [
          "tasks"
        ],
        "type": "object",
        "properties": {
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectorTaskResponse"
            }
          }
        }
      },
      "TaskProgressRequest": {
        "required": [
          "timeElapsed",
          "percentageComplete",
          "attributes"
        ],
        "type": "object",
        "properties": {
          "timeElapsed": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "percentageComplete": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "attributes": {
            "$ref": "#/components/schemas/JsonElement"
          }
        }
      },
      "TaskProgressResponse": {
        "required": [
          "taskProgressId"
        ],
        "type": "object",
        "properties": {
          "taskProgressId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "TaskSettingsResponse": {
        "required": [
          "taskId",
          "scanId",
          "scanConfigId",
          "settings",
          "authMethod"
        ],
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string",
            "format": "uuid"
          },
          "scanId": {
            "type": "string",
            "format": "uuid"
          },
          "scanConfigId": {
            "type": "string",
            "format": "uuid"
          },
          "settings": {
            "type": "string"
          },
          "authMethod": {
            "type": "string"
          }
        }
      },
      "UpdateTaskStatusRequest": {
        "required": [
          "status"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Tasks"
    },
    {
      "name": "Uploads"
    }
  ]
}